chore: Always run init to create symlink (main)
Force `make init` to run even if the symlink exists. This
ensures that the symlink is always created correctly,
preventing potential build issues.

+++ Makefile
@@ -1,8 +1,8 @@
 KEYMAP_DIR = "qmk_firmware/keyboards/crkbd/keymaps/n"
 
 build:
-	@[ -L $(KEYMAP_DIR) ] || make init
+	make init
 	cd qmk_firmware && qmk compile -kb crkbd -km n
 
 init:
-	ln -s $(PWD)/n $(KEYMAP_DIR)
+	@[ -L $(KEYMAP_DIR) ] || ln -s $(PWD)/n $(KEYMAP_DIR)
